Fix some includes to optimize build speed (#45393)
