+++ ydb/core/filestore/core/filestore.h
@@ -2,7 +2,7 @@
 
 #include <ydb/library/actors/core/actor.h>
 #include <ydb/library/actors/core/actorid.h>
-#include <ydb/core/base/blobstorage.h>
+#include <ydb/core/base/events.h>
 #include <ydb/core/protos/filestore_config.pb.h>
 
 namespace NKikimr {
