+++ ydb/core/tablet/tablet_counters_aggregator.h
@@ -8,7 +8,6 @@
 #include <ydb/library/actors/core/actor.h>
 #include <ydb/library/actors/core/event.h>
 
-#include <ydb/core/base/blobstorage.h>
 #include <ydb/core/protos/labeled_counters.pb.h>
 #include <ydb/core/protos/tablet_counters_aggregator.pb.h>
 #include <ydb/core/sys_view/common/events.h>
@@ -46,13 +45,13 @@ struct TEvTabletCounters {
     struct TEvTabletAddCounters : public TEventLocal<TEvTabletAddCounters, EvTabletAddCounters> {
         //
         const ui64 TabletID;
-        const TTabletTypes::EType TabletType;
+        const NKikimrTabletBase::TTabletTypes::EType TabletType;
         const TPathId TenantPathId;
         TAutoPtr<TTabletCountersBase> ExecutorCounters;
         TAutoPtr<TTabletCountersBase> AppCounters;
         TIntrusivePtr<TInFlightCookie> InFlightCounter;     // Used to detect when previous event has been consumed by the aggregator
 
-        TEvTabletAddCounters(TIntrusivePtr<TInFlightCookie> inFlightCounter, ui64 tabletID, TTabletTypes::EType tabletType, TPathId tenantPathId,
+        TEvTabletAddCounters(TIntrusivePtr<TInFlightCookie> inFlightCounter, ui64 tabletID, NKikimrTabletBase::TTabletTypes::EType tabletType, TPathId tenantPathId,
             TAutoPtr<TTabletCountersBase> executorCounters, TAutoPtr<TTabletCountersBase> appCounters)
             : TabletID(tabletID)
             , TabletType(tabletType)
@@ -65,10 +64,10 @@ struct TEvTabletCounters {
 
     struct TEvTabletAddLabeledCounters : public TEventLocal<TEvTabletAddLabeledCounters, EvTabletAddLabeledCounters> {
         const ui64 TabletID;
-        const TTabletTypes::EType TabletType;
+        const NKikimrTabletBase::TTabletTypes::EType TabletType;
         TAutoPtr<TTabletLabeledCountersBase> LabeledCounters;
         TIntrusivePtr<TInFlightCookie> InFlightCounter;     // Used to detect when previous event has been consumed by the aggregator
-        TEvTabletAddLabeledCounters(TIntrusivePtr<TInFlightCookie> inFlightCounter, ui64 tabletID, TTabletTypes::EType tabletType, TAutoPtr<TTabletLabeledCountersBase> labeledCounters)
+        TEvTabletAddLabeledCounters(TIntrusivePtr<TInFlightCookie> inFlightCounter, ui64 tabletID, NKikimrTabletBase::TTabletTypes::EType tabletType, TAutoPtr<TTabletLabeledCountersBase> labeledCounters)
             : TabletID(tabletID)
             , TabletType(tabletType)
             , LabeledCounters(labeledCounters)
@@ -80,10 +79,10 @@ struct TEvTabletCounters {
     struct TEvTabletCountersForgetTablet : public TEventLocal<TEvTabletCountersForgetTablet, EvTabletCountersForgetTablet> {
         //
         const ui64 TabletID;
-        const TTabletTypes::EType TabletType;
+        const NKikimrTabletBase::TTabletTypes::EType TabletType;
         const TPathId TenantPathId;
 
-        TEvTabletCountersForgetTablet(ui64 tabletID, TTabletTypes::EType tabletType, TPathId tenantPathId)
+        TEvTabletCountersForgetTablet(ui64 tabletID, NKikimrTabletBase::TTabletTypes::EType tabletType, TPathId tenantPathId)
             : TabletID(tabletID)
             , TabletType(tabletType)
             , TenantPathId(tenantPathId)
@@ -126,7 +125,7 @@ struct TTabletLabeledCountersResponseContext {
 };
 
 ////////////////////////////////////////////
-void TabletCountersForgetTablet(ui64 tabletId, TTabletTypes::EType tabletType, TPathId tenantPathId, bool follower, TActorIdentity identity);
+void TabletCountersForgetTablet(ui64 tabletId, NKikimrTabletBase::TTabletTypes::EType tabletType, TPathId tenantPathId, bool follower, TActorIdentity identity);
 
 TStringBuf GetHistogramAggregateSimpleName(TStringBuf name);
 bool IsHistogramAggregateSimpleName(TStringBuf name);
@@ -145,14 +144,14 @@ IActor* CreateTabletCountersAggregator(bool follower);
 //will create actor that aggregate LabeledCounters from all nodes and reports them as TEvTabletLabeledCountersResponse to parentActor
 TActorId CreateClusterLabeledCountersAggregator(
         const TActorId& parentActor,
-        TTabletTypes::EType tabletType,
+        NKikimrTabletBase::TTabletTypes::EType tabletType,
         const TActorContext& ctx,
         ui32 version = 1,
         const TString& group = TString(), const ui32 TotalWorkersCount = WORKERS_COUNT);
 
 IActor* CreateClusterLabeledCountersAggregatorActor(
         const TActorId& parentActor,
-        TTabletTypes::EType tabletType,
+        NKikimrTabletBase::TTabletTypes::EType tabletType,
         ui32 version = 1,
         const TString& group = TString(), const ui32 TotalWorkersCount = WORKERS_COUNT);
