+++ ydb/core/tablet_flat/ut/ut_shared_sausagecache_actor.cpp
@@ -9,6 +9,7 @@
 #include <ydb/core/testlib/basics/appdata.h>
 #include <ydb/library/actors/testlib/test_runtime.h>
 #include <library/cpp/testing/unittest/registar.h>
+#include <ydb/core/protos/bootstrap.pb.h>
 
 namespace NKikimr::NSharedCache {
 using namespace NActors;
