+++ ydb/core/tx/datashard/datashard.cpp
@@ -5113,4 +5113,13 @@ size_t TEvDataShard::TEvReadResult::GetDataSizeEstimate() const {
     return 0;
 }
 
+void TEvDataShard::TEvProposeTransactionResult::SetStepOrderId(const std::pair<ui64, ui64>& stepOrderId) {
+    Record.SetStep(stepOrderId.first);
+    Record.SetOrderId(stepOrderId.second);
+    // Note: this method is used by schema operations where stepOrderId == commitVersion
+    auto* commitVersion = Record.MutableCommitVersion();
+    commitVersion->SetStep(stepOrderId.first);
+    commitVersion->SetTxId(stepOrderId.second);
+}
+
 } // NKikimr
