+++ ydb/core/tx/schemeshard/schemeshard.h
@@ -17,6 +17,12 @@
 #include <library/cpp/object_factory/object_factory.h>
 
 namespace NKikimr {
+
+// Only used below as a pointer in a factory-function declaration; a
+// forward declaration avoids pulling the heavy ydb/core/base/blobstorage.h
+// into this widely-included header.
+class TTabletStorageInfo;
+
 namespace NSchemeShard {
 
 static constexpr ui64 RootSchemeShardId = 0;
